Shop Update
I reworked bits of the shop so that items would be on the canvas and work as buttons. The shop is now more connected to the old script; any changes to the old shop script that I made are accompanied by comments with "Cerulean", so you can search for them with that.