New Shop Update 1.1
I basically added functionality with the new input system, and while it works, there are some bugs that seem to be the result of a separate instance of the input system doing it's own thing in the background, and I don't know where it's coming from or exactly how it's modfying item selection values.

For example, you can select an item before the items fade into view, however, when I tested this with debug logs, I realized it wasn't coming from my script. Furthermore, I believe the two instances can go out of sync at times as when I button-mash, the item that's currently selected doesn't always match what the game is showing me.